Fix having the same stream multiple times
Since browse goes off the most watched streams, the ordering changes given
changes in viewership. So it's possible that a stream gets returned multiple
times because it has fallen lower as viewers left. I expected this to be fixed
thanks to the `cursor` we get from twitch, however that does not seem to be the
case.

Therefor we make sure to update existing entries if we receive a stream entry
which we already have before.