Properly configure public url when building
Currently the application is being deployed under `/seeker`. By default,
vue sets the base path to `/`, making all assets be requested using that
path. However, since we're using github pages to deploy, it will be
deployed under `/seeker` and all resources need to use that as a base.
We configure this in the cli configuration file which is used in the
build process.