Use single state change for adding name mappings
When loading streams, we add a mapping to the ID for each of them. This is done
after receiving them as a whole making it rather unnecessary to create separate
store commits for each of them. It also slows down the devtools since they have
to recollect a lot of changes to the store.