Only persist games, streams and tags we need a name for
Currently, we persist all tags, games and stream name mappings we load over the
course of using it. This causes the local storage to fill up eventually, given
the amount of information is contained in these. I initially thought about
stripping unnecessary information from them (such as picture or localization
info) but decided that this was not the right way to go as it would prevent
adding localization support in the case of tags or not allowing us to show the
image in the future without requesting them again. By only storing the ones we
display, we make sure that these are already loaded, no matter what streams show
 up.