Properly load tags
We didn't quite load tags properly, specifically if we had more than one hundred
 tags. In that case, it would not work properly because twitch only accepts 100
tags to be passed to the api at a time and we'd throw all of them at it. We only
 have this issue with tags as all the others are effectively guaranteed to only
be requested in chunks smaller than 100.

We fix this by chunking the tags into pages of 100 which is the maximum.