Move ignore action to an overlay on top of stream thumbnail
When we moved away from the cards, we also made the the ignore button to pop in
and out if the user hovers over the stream. This was rather annoying as it also
moved the content below and was rather hard to hit. Instead, it's now a small
button on top of the stream thumbnail and it's using an icon instead to make it
use less space.