Change the remove action of already filtered tags
To remove a tag/stream/game that was already filtered, there was a small X added
 to each badge. This was not really that visible and we didn't need to display
that all the time. So instead of making it just visible while hover, it got
removed entirely and clicking on the badge will work instead. This makes it a
bit more intuitive as well.