Use github actions to build upon push
Currently, I manually build the changes and push them onto the gh-pages
branch which sometimes contains unwanted changes and can be forgotten.
Especially if outside contributions come in, it would be great if things
just automatically get published without my doing.

To allow that, the twitch api key is now defined as a constant which is
replaced by webpack. Therefor, it no longer needs to be replaced, but
instead be supplied as an environment variable when calling npm build or
serve.