Require login for functionality
Twitch now requires authorization for many more APIs, including the stream
listing functionality. One could use an app token for this purpose but they are
not available for client only applications. As a result, we have to require the
users to login before using the application. This currently does not handle code
 expiry _while_ using the application but it should be detected on reload.

One thing still marked as TODO is that the streams endpoint can return tags that
 are not applied directly to the stream but are implied from the category the
stream is in. However, when requesting more data for such a tag, the tags
endpoint does not return any data. Only when the specific stream and its tags
are requested, it will return the data for the tag. As a result, this could
accidentally request quite a lot of data that I currently would like to avoid.
Some streams are therefore missing some tags that are visible on twitch.

This also cleaned up some markup and styles in the meantime.